+++ src/connectedvmware/azext_connectedvmware/_params.py
@@ -351,6 +351,10 @@ def load_arguments(self, _):
             'password', options_list=['--password'],
             help="Username password credentials to use for connecting to the VM.",
         )
+        c.argument(
+            'https_proxy', options_list=['--https-proxy'],
+            help="HTTPS proxy server url for the VM.",
+        )
 
     with self.argument_context('connectedvmware vm guest-agent show') as c:
         c.argument(